requirements.txt
# ══════════════════════════════════════════════
# GB-BRAIN — Python Dependencies
# Install: pip install -r requirements.txt
# ══════════════════════════════════════════════

# ── Core ─────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0

# ── Data Sources ─────────────────────────────
yfinance>=0.2.36
python-binance>=1.0.19

# ── Technical Analysis ───────────────────────
ta>=0.11.0

# ── Backtesting ──────────────────────────────
backtesting>=0.3.3

# ── Excel Reports ────────────────────────────
openpyxl>=3.1.0

# ── Web Server (webhook + dashboard) ─────────
flask>=3.0.0

# ── Telegram ─────────────────────────────────
python-telegram-bot>=20.0

# ── OANDA ────────────────────────────────────
oandapyV20>=0.7.2

# ── Alpaca ───────────────────────────────────
alpaca-trade-api>=3.0.0
